requirements.txt
langchain-cohere
pinecone
python-decouple
streamlit
flask